requirements.txt
alabaster==1.0.0
asgiref==3.8.1
babel==2.16.0
certifi==2024.8.30
colorama==0.4.6
docutils==0.21.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.4
MarkupSafe==3.0.0
packaging==24.1
Pygments==2.18.0
requests==2.32.3
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinx-rtd-theme==3.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tzdata==2024.2
urllib3==2.2.3
Pillow==11.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.3
django-allauth==65.2.0
django-requests==0.0.2
idna==3.10
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
